Try moving AssignState
