+++ parlai/mturk/core/agents.py
@@ -14,160 +14,14 @@ import parlai.mturk.core.data_model as data_model
 import parlai.mturk.core.shared_utils as shared_utils
 
 # Special act messages for failure states
+from parlai.mturk.core.shared_utils import AssignState
+
 MTURK_DISCONNECT_MESSAGE = '[DISCONNECT]'  # Turker disconnected from conv
 TIMEOUT_MESSAGE = '[TIMEOUT]'  # the Turker did not respond but didn't return
 RETURN_MESSAGE = '[RETURNED]'  # the Turker returned the HIT
 
 # TODO move time management into another class, this way we can handle it
 # relative to heartbeats. This will come with more thorough testing.
-
-
-class AssignState:
-    """
-    Class for holding state information about an assignment currently claimed by an
-    agent.
-    """
-
-    # Possible Assignment Status Values
-    STATUS_NONE = 'none'
-    STATUS_ONBOARDING = 'onboarding'
-    STATUS_WAITING = 'waiting'
-    STATUS_IN_TASK = 'in task'
-    STATUS_DONE = 'done'
-    STATUS_DISCONNECT = 'disconnect'
-    STATUS_PARTNER_DISCONNECT = 'partner disconnect'
-    STATUS_PARTNER_DISCONNECT_EARLY = 'partner disconnect early'
-    STATUS_EXPIRED = 'expired'
-    STATUS_RETURNED = 'returned'
-
-    def __init__(self, status=None):
-        """
-        Create an AssignState to track the state of an agent's assignment.
-        """
-        if status is None:
-            status = self.STATUS_NONE
-        self.status = status
-        self.messages = []
-        self.last_command = None
-        self.message_ids = []
-
-    def clear_messages(self):
-        self.messages = []
-        self.message_ids = []
-        self.last_command = None
-
-    def append_message(self, message):
-        """
-        Appends a message to the list of messages, ensures that it is not a duplicate
-        message.
-        """
-        if message['message_id'] in self.message_ids:
-            return
-        self.message_ids.append(message['message_id'])
-        self.messages.append(message)
-
-    def set_last_command(self, command):
-        self.last_command = command
-
-    def get_last_command(self):
-        return self.last_command
-
-    def get_messages(self):
-        return self.messages
-
-    def set_status(self, status):
-        """
-        Set the status of this agent on the task.
-        """
-        # TODO log to db
-        self.status = status
-
-    def get_status(self):
-        """
-        Get the status of this agent on its task.
-        """
-        # TODO retrieve from db if not set
-        return self.status
-
-    def is_final(self):
-        """
-        Return True if the assignment is in a final status that can no longer be acted
-        on.
-        """
-        return (
-            self.status == self.STATUS_DISCONNECT
-            or self.status == self.STATUS_DONE
-            or self.status == self.STATUS_PARTNER_DISCONNECT
-            or self.status == self.STATUS_PARTNER_DISCONNECT_EARLY
-            or self.status == self.STATUS_RETURNED
-            or self.status == self.STATUS_EXPIRED
-        )
-
-    def get_inactive_command_text(self):
-        """
-        Get appropriate inactive command and text to respond to a reconnect given the
-        current assignment state.
-
-        returns text, command
-        """
-        command = data_model.COMMAND_INACTIVE_HIT
-        text = None
-        if self.status == self.STATUS_DISCONNECT:
-            text = (
-                'You disconnected in the middle of this HIT and were '
-                'marked as inactive. As these HITs often require real-'
-                'time interaction, it is no longer available for '
-                'completion. Please return this HIT and accept a new one '
-                'if you would like to try again.'
-            )
-        elif self.status == self.STATUS_DONE:
-            command = data_model.COMMAND_INACTIVE_DONE
-            text = (
-                'You disconnected after completing this HIT without '
-                'marking it as completed. Please press the done button '
-                'below to finish the HIT.'
-            )
-        elif self.status == self.STATUS_EXPIRED:
-            text = (
-                'You disconnected in the middle of this HIT and the '
-                'HIT expired before you reconnected. It is no longer '
-                'available for completion. Please return this HIT and '
-                'accept a new one if you would like to try again.'
-            )
-        elif self.status == self.STATUS_PARTNER_DISCONNECT:
-            command = data_model.COMMAND_INACTIVE_DONE
-            text = (
-                'One of your partners disconnected in the middle of the '
-                'HIT. We won\'t penalize you for their disconnect, so '
-                'please use the button below to mark the HIT as complete.'
-            )
-        elif self.status == self.STATUS_PARTNER_DISCONNECT_EARLY:
-            command = data_model.COMMAND_INACTIVE_HIT
-            text = (
-                'One of your partners disconnected in the middle of the '
-                'HIT. We won\'t penalize you for their disconnect, but you'
-                ' did not complete enough of the task to submit the HIT. '
-                'Please return this HIT and accept a new one if you would '
-                'like to try again.'
-            )
-        elif self.status == self.STATUS_RETURNED:
-            text = (
-                'You disconnected from this HIT and then returned '
-                'it. As we have marked the HIT as returned, it is no '
-                'longer available for completion. Please accept a new '
-                'HIT if you would like to try again'
-            )
-        else:
-            # We shouldn't be getting an inactive command for the other
-            # states so consider this a server error
-            text = (
-                'Our server was unable to handle your reconnect properly '
-                'and thus this HIT no longer seems available for '
-                'completion. Please try to connect again or return this '
-                'HIT and accept a new one.'
-            )
-
-        return text, command
 
 
 class MTurkAgent(Agent):
