+++ parlai/mturk/core/legacy_2018/mturk_manager.py
@@ -14,7 +14,7 @@ import uuid
 import errno
 import requests
 
-from parlai.mturk.core.agents import AssignState
+from parlai.mturk.core.shared_utils import AssignState
 from parlai.mturk.core.socket_manager import Packet, SocketManager, StaticSocketManager
 from parlai.mturk.core.worker_manager import WorkerManager
 from parlai.mturk.core.mturk_data_handler import MTurkDataHandler
