+++ parlai/mturk/core/legacy_2018/test/test_full_system.py
@@ -10,7 +10,7 @@ import uuid
 import os
 from unittest import mock
 from parlai.mturk.core.socket_manager import Packet, SocketManager
-from parlai.mturk.core.agents import AssignState
+from parlai.mturk.core.shared_utils import AssignState
 from parlai.mturk.core.mturk_manager import MTurkManager
 from parlai.core.params import ParlaiParser
