+++ parlai/mturk/core/test/test_worker_manager.py
@@ -8,7 +8,8 @@ import unittest
 import os
 from unittest import mock
 from parlai.mturk.core.worker_manager import WorkerManager, WorkerState
-from parlai.mturk.core.agents import MTurkAgent, AssignState
+from parlai.mturk.core.agents import MTurkAgent
+from parlai.mturk.core.shared_utils import AssignState
 from parlai.mturk.core.mturk_manager import MTurkManager
 from parlai.core.params import ParlaiParser
