+++ parlai/mturk/webapp/run_mocks/mock_turk_manager.py
@@ -10,7 +10,7 @@ import threading
 import time
 import uuid
 
-from parlai.mturk.core.agents import AssignState
+from parlai.mturk.core.shared_utils import AssignState
 from parlai.mturk.core.socket_manager import Packet
 from parlai.mturk.webapp.run_mocks.mock_turk_agent import MockTurkAgent
 import parlai.mturk.core.data_model as data_model
